fix some linters
