+++ modules/render.js
@@ -1,13 +1,21 @@
-import { Book, booksArr, updateBooksArr } from './book.js';
+import Book from './book.js';
 
+let booksArr = [];
+
+const titleInput = document.querySelector('.title-input');
+const authorInput = document.querySelector('.author-input');
+const addBookBtn = document.querySelector('.add-book-btn');
 const bookContainer = document.querySelector('.books-container');
 
 const render = () => {
   const book = new Book();
   bookContainer.innerHTML = null;
   if (localStorage.getItem('books')) {
-    updateBooksArr(JSON.parse(localStorage.getItem('books')));
+    booksArr = JSON.parse(localStorage.getItem('books'));
   }
+  const removeBook = (index) => {
+    book.removeBook(index, booksArr);
+  };
   for (let i = 0; i < booksArr.length; i += 1) {
     const bookDiv = document.createElement('div');
     const bookWraper = document.createElement('div');
@@ -30,9 +38,15 @@ const render = () => {
     bookContainer.appendChild(bookDiv);
     removeBookBtn.addEventListener('click', () => {
       bookDiv.remove();
-      book.removeBook(i);
-      });
+      removeBook(i);
+    });
   }
 };
 
-export { render }
+addBookBtn.addEventListener('click', () => {
+  const book = new Book(titleInput.value, authorInput.value);
+  book.add(book, booksArr);
+  render();
+});
+
+export default render;
